store rewrite: delete the legacy implementation
The rewrite now serves every public store form, so legacy goes:
reconcile/projection/optimistic modules removed, store.ts gutted to
shared machinery (~1600 → ~570 lines: symbols, raw-marking,
isWrappable, write-override, next-only affects walk, StoreNode as a
structural view of StoreNextTarget), utils.ts to merge/omit, the
transitional dispatchers and test shims removed — store/index exports
next directly. createWriteTraps lives in next/projection; next's
optimistic hooks are self-contained (engine + own clear + blocked
half). Orphan sweep: storeLookup, symbolKeyedRecords (+ skipped
legacy-pinned test), registerTransientStoreNode, mergedOverlay.
Suite 1253 green; full monorepo build + tests green; in-process dbmon
bench in family. Size vs main (gzip): full 24162 → 23259 (−3.7%),
store attribution 15572 → 14803.
