+++ packages/solid-signals/src/store/next/projection.ts
@@ -23,10 +23,12 @@ import {
   type Computed,
   type Refreshable
 } from "../../core/index.js";
-import { createWriteTraps } from "../projection.js";
+
+import { projectionWriteActive, setProjectionWriteActive } from "../../core/scheduler.js";
 import {
   $TARGET,
   markRawIngest,
+  setWriteOverride,
   STORE_VALUE,
   type NoFn,
   type ProjectionOptions,
@@ -35,6 +37,70 @@ import {
 import { reconcileNextState } from "./reconcile.js";
 import { storeSetterNext, wrapNext } from "./store.js";
 import type { StoreNextFamily } from "./target.js";
+
+function createWriteTraps(isActive?: () => boolean, onDraftWrite?: () => void): ProxyHandler<any> {
+  // Save/restore, never hard-reset: the draft can be driven from inside an
+  // enclosing authoritative-write scope (next-store optimistic derives), and
+  // a hard `false` would clobber it mid-derive.
+  const traps: ProxyHandler<any> = {
+    get(_, prop) {
+      let value;
+      const was = projectionWriteActive;
+      setWriteOverride(true);
+      setProjectionWriteActive(true);
+      try {
+        value = _[prop];
+      } finally {
+        setWriteOverride(false);
+        setProjectionWriteActive(was);
+      }
+      if (prop === $TARGET) return value;
+      return typeof value === "object" && value !== null ? new Proxy(value, traps) : value;
+    },
+    has(_, prop) {
+      let value;
+      const was = projectionWriteActive;
+      setWriteOverride(true);
+      setProjectionWriteActive(true);
+      try {
+        value = prop in _;
+      } finally {
+        setWriteOverride(false);
+        setProjectionWriteActive(was);
+      }
+      return value;
+    },
+    set(_, prop, value) {
+      if (isActive && !isActive()) return true;
+      const was = projectionWriteActive;
+      setWriteOverride(true);
+      setProjectionWriteActive(true);
+      try {
+        _[prop] = value;
+        onDraftWrite?.();
+      } finally {
+        setWriteOverride(false);
+        setProjectionWriteActive(was);
+      }
+      return true;
+    },
+    deleteProperty(_, prop) {
+      if (isActive && !isActive()) return true;
+      const was = projectionWriteActive;
+      setWriteOverride(true);
+      setProjectionWriteActive(true);
+      try {
+        delete _[prop];
+        onDraftWrite?.();
+      } finally {
+        setWriteOverride(false);
+        setProjectionWriteActive(was);
+      }
+      return true;
+    }
+  };
+  return traps;
+}
 
 export function createProjectionNextInternal<T extends object = {}>(
   fn: (draft: T) => void | T | Promise<void | T> | AsyncIterable<void | T>,
