+++ packages/solid-signals/src/store/utils.ts
@@ -1,191 +1,7 @@
 import { pendingCheckActive } from "../core/core.js";
 import { SUPPORTS_PROXY } from "../core/index.js";
 import { createMemo } from "../signals.js";
-import {
-  $DELETED,
-  $PROXY,
-  $TARGET,
-  $TRACK,
-  getKeys,
-  getStoreKeys,
-  getStoreSymbols,
-  getPropertyDescriptor,
-  isWrappable,
-  mergedOverlay,
-  ownEnumerableKeys,
-  STORE_LOOKUP,
-  STORE_VALUE,
-  storeLookup,
-  lookupTarget,
-  isRawValue,
-  rawValuesUsed,
-  trackSelf,
-  witnessAffectsMark,
-  wrap,
-  type StoreNode
-} from "./store.js";
-
-function snapshotImpl<T>(
-  item: any,
-  track: boolean,
-  map?: Map<unknown, unknown>,
-  lookup?: WeakMap<any, any>
-): T {
-  let target: StoreNode | undefined, isArray, override, result, unwrapped, v;
-  if (!isWrappable(item)) return item;
-  if (map && map.has(item)) return map.get(item) as T;
-  if (!map) map = new Map();
-  if ((target = item[$TARGET] || lookupTarget(item, lookup))) {
-    if (track) {
-      trackSelf(target, $TRACK);
-      // A tracked walk reads THROUGH the record without touching the proxy
-      // traps — witness the record's affects() channel like a trap read would.
-      if (pendingCheckActive) witnessAffectsMark(target);
-    }
-    override = mergedOverlay(target);
-    // A derived store's STORE_VALUE is the inner store's live proxy
-    // (store-in-store: createOptimisticStore/createProjection over a store).
-    // Without an overlay of its own, the fast path below would map to — and
-    // could return — that proxy verbatim. Recurse instead so the
-    // inner store's own target branch unwraps it (chains of any depth).
-    // With an overlay, a fresh `result` is always built, so the walk over the
-    // inner proxy already copies plain values.
-    if (!override && (target[STORE_VALUE] as StoreNode)[$TARGET]) {
-      unwrapped = snapshotImpl(target[STORE_VALUE], track, map, lookup);
-      map.set(item, unwrapped);
-      return unwrapped as T;
-    }
-    isArray = Array.isArray(target[STORE_VALUE]);
-    map.set(
-      item,
-      override
-        ? (result = isArray ? [] : (Object.create(Object.getPrototypeOf(target[STORE_VALUE])) as T))
-        : target[STORE_VALUE]
-    );
-    item = target[STORE_VALUE];
-    lookup = target[STORE_LOOKUP] ?? storeLookup;
-  } else {
-    isArray = Array.isArray(item);
-    map.set(item, item);
-  }
-  if (isArray) {
-    const len = override?.length ?? item.length;
-    for (let i = 0; i < len; i++) {
-      v = override && i in override ? override[i] : item[i];
-      if (v === $DELETED) continue;
-      if (track && isWrappable(v) && !(rawValuesUsed && isRawValue(v))) wrap(v, target);
-      if ((unwrapped = snapshotImpl(v, track, map, lookup)) !== v || result) {
-        if (!result) map.set(item, (result = [...item]));
-        result[i] = unwrapped;
-      }
-    }
-    // Enumerate array symbols separately to avoid scanning indices twice.
-    // Spread copies omit symbols, so assign them after the numeric walk.
-    const symbols = lookup ? getStoreSymbols(item, override) : [];
-    for (let i = 0, l = symbols.length; i < l; i++) {
-      const prop = symbols[i];
-      const desc = getPropertyDescriptor(item, override, prop);
-      if (!desc || desc.get) continue;
-      v = override && prop in override ? override[prop] : item[prop];
-      if (track && isWrappable(v) && !(rawValuesUsed && isRawValue(v))) wrap(v, target);
-      unwrapped = snapshotImpl(v, track, map, lookup);
-      if (unwrapped !== v || result) {
-        if (!result) map.set(item, (result = Object.assign([...item], item)));
-        result[prop] = unwrapped;
-      }
-    }
-    // Deleted trailing slots are skipped above, so restore length to preserve
-    // holes instead of truncating the copy (#2846) — mirrors unwrapStoreValue.
-    if (result) result.length = len;
-  } else if (!override) {
-    // Specialized walk for the common no-overlay case (from #2756): the own
-    // descriptor gives the value directly, so each property is read once with
-    // no overlay membership checks.
-    // A lookup means this object belongs to an immutable store backing tree,
-    // even if that nested value has not needed its own proxy yet.
-    const keys = lookup ? getStoreKeys(item, undefined) : getKeys(item, undefined);
-    for (let i = 0, l = keys.length; i < l; i++) {
-      const prop = keys[i];
-      const desc = Object.getOwnPropertyDescriptor(item, prop)!;
-      if (desc.get) continue;
-      v = desc.value;
-      if (track && isWrappable(v) && !(rawValuesUsed && isRawValue(v))) wrap(v, target);
-      if ((unwrapped = snapshotImpl(v, track, map, lookup)) !== v || result) {
-        if (!result) {
-          result = Object.create(Object.getPrototypeOf(item)) as Record<PropertyKey, any>;
-          Object.assign(result, item);
-        }
-        result[prop] = unwrapped;
-      }
-    }
-  } else {
-    // An override only exists on a store record, and the target branch above
-    // always set `lookup` alongside it — so this branch is always store-keyed.
-    const keys = getStoreKeys(item, override);
-    for (let i = 0, l = keys.length; i < l; i++) {
-      let prop = keys[i];
-      const desc = getPropertyDescriptor(item, override, prop)!;
-      if (desc.get) continue;
-      v = prop in override ? override[prop] : item[prop];
-      if (track && isWrappable(v) && !(rawValuesUsed && isRawValue(v))) wrap(v, target);
-      if ((unwrapped = snapshotImpl(v, track, map, lookup)) !== item[prop] || result) {
-        if (!result) {
-          result = Object.create(Object.getPrototypeOf(item)) as Record<PropertyKey, any>;
-          Object.assign(result, item);
-        }
-        result[prop] = unwrapped;
-      }
-    }
-  }
-  return result || item;
-}
-
-/**
- * Returns a plain (non-proxy, non-reactive) deep copy of a store value.
- * Reading via `snapshot` does **not** subscribe to changes — use this when you
- * need to hand a stable plain object to non-reactive code (logging,
- * serialization, structured-clone, network payloads, etc.).
- *
- * Returns the original object identity for any sub-tree that wasn't modified
- * relative to the proxy's underlying source.
- *
- * @example
- * ```ts
- * const [state] = createStore({ user: { name: "Ada" }, todos: [] });
- *
- * console.log(JSON.stringify(snapshot(state))); // safe, non-reactive copy
- * ```
- */
-export function snapshot<T>(item: T): T;
-export function snapshot<T>(item: T, map?: Map<unknown, unknown>, lookup?: WeakMap<any, any>): T;
-export function snapshot<T>(item: any, map?: Map<unknown, unknown>, lookup?: WeakMap<any, any>): T {
-  return snapshotImpl(item, false, map, lookup);
-}
-
-/**
- * Returns a plain (non-proxy) deep copy **and** subscribes the current
- * tracking scope to every nested change in the source store. Any write
- * anywhere in the subtree invalidates the consumer.
- *
- * Use this when you need plain data inside a reactive scope and want to
- * react to deep mutations (e.g. passing a snapshot to `reconcile()` or to a
- * memo that should rerun on any nested change). For most read paths, prefer
- * direct property access — Solid stores already track per-property reads
- * with no `deep()` wrapper needed.
- *
- * @example
- * ```ts
- * const [state] = createStore({ a: { b: { c: 1 } } });
- *
- * createEffect(
- *   () => deep(state),                  // reruns on any nested change
- *   plain => sendToWorker(plain)        // worker gets a non-proxy copy
- * );
- * ```
- */
-export function deep<T extends object>(store: T): T {
-  return snapshotImpl(store, true) as T;
-}
+import { $PROXY, ownEnumerableKeys } from "./store.js";
 
 function trueFn() {
   return true;
